feat(document): Add preview for uploaded documents
On the list of files :
If a document has no preview url attached, display the original document
Else display the preview url attached and no longer the original file.

When a user open the preview :
If a watermark document has been created display the entire document.
Else, display the original document.